+++ src/index.ts
@@ -1,12 +1,13 @@
 import { Stream, SinkStream } from "./stream";
 import { Behavior, SinkBehavior } from "./behavior";
 
+export * from "./common";
 export * from "./behavior";
 export * from "./stream";
 export * from "./future";
-// export * from "./now";
-// export * from "./dom";
-// export * from "./placeholder";
+export * from "./now";
+export * from "./dom";
+export * from "./placeholder";
 
 /**
  * Map a function over a behavior or stream. This means that if at some point in
